requirements.txt
hidapi
pycryptodome
numpy
scipy
pandas
matplotlib
stable-baselines3
gym
djitellopy
scikit-learn